pyproject.toml
[project]
name = "omniview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.0.0",
    "langgraph>=0.4.5",
    "pydantic-settings>=2.9.1",
    "langchain-openai>=0.3.22",
    "langchain-core>=0.3.65",
    "langgraph-cli[inmem]>=0.2.10",
    "langsmith>=0.3.43",
    "requests>=2.32.3",
]
